+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changes are grouped as follows
 - Separate read/write fields on data classes
 
 ## [Unreleased]
+- Sequences data insert no longer takes a default value for columns.
 
 ## [1.1.12] - 2019-10-01
 ### Fixed